requirements/base.txt
asgiref==3.5.0
attrs==21.4.0
Babel==2.9.1
backports.zoneinfo==0.2.1
Django==3.2.12
django-money==2.1.1
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
importlib-metadata==4.11.3
iniconfig==1.1.1
ipython==7.34.0
model-bakery==1.4.0
packaging==21.3
phonenumbers==8.12.44
pluggy==1.0.0
py==1.11.0
py-moneyed==1.2
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
zipp==3.7.0
